+++ src/collective/cover/tiles/carousel.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-
 from collective.cover import _
 from collective.cover.interfaces import ITileEditForm
 from collective.cover.tiles.list import IListTile
@@ -12,28 +11,17 @@ from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from zope import schema
 from zope.interface import implements
 
-# autoplay feature is enabled in view mode only
-INIT_JS = """$(function() {{
-    Galleria.loadTheme('++resource++collective.cover/galleria-theme/galleria.cover_theme.js');
-    Galleria.run('#galleria-{0}');
-
-    var options = {{ height: 1 }};
-    if ($('body').hasClass('template-view')) {{
-        options.autoplay = {1};
-    }}
-    Galleria.configure(options);
-}});
-"""
-
 
 class ICarouselTile(IListTile):
-    """A carousel based on the Galleria JavaScript image gallery framework.
-    """
 
+    """A carousel based on the Cycle2 slideshow plugin for jQuery."""
+
+    form.omitted('autoplay')
+    form.no_omit(ITileEditForm, 'autoplay')
     autoplay = schema.Bool(
         title=_(u'Auto play'),
+        default=False,
         required=False,
-        default=True,
     )
 
     form.no_omit(ITileEditForm, 'uuids')
@@ -48,8 +36,9 @@ class ICarouselTile(IListTile):
 
 class CarouselTile(ListTile):
 
-    implements(ICarouselTile)
+    """A carousel based on the Cycle2 slideshow plugin for jQuery."""
 
+    implements(ICarouselTile)
     index = ViewPageTemplateFile('templates/carousel.pt')
     is_configurable = True
     is_editable = True
@@ -57,12 +46,7 @@ class CarouselTile(ListTile):
 
     def populate_with_object(self, obj):
         super(CarouselTile, self).populate_with_object(obj)  # check permission
-        try:
-            scale = obj.restrictedTraverse('@@images').scale('image')
-        except:
-            scale = None
-        if not scale:
-            return
+
         self.set_limit()
         uuid = IUUID(obj, None)
         data_mgr = ITileDataManager(self)
@@ -80,17 +64,8 @@ class CarouselTile(ListTile):
             old_data['uuids'] = [uuid]
         data_mgr.set(old_data)
 
-    def autoplay(self):
-        if self.data['autoplay'] is None:
-            return True  # default value
-
-        return self.data['autoplay']
-
-    def init_js(self):
-        if self.is_empty():
-            # Galleria will display scary error messages when it
-            # cannot find its <div>.  So don't start galleria unless
-            # the <div> is there and has some items in it.
-            return ''
-
-        return INIT_JS.format(self.id, str(self.autoplay()).lower())
+    @property
+    def paused(self):
+        """Return True if the carousel will begin in a paused state."""
+        paused = not self.data.get('autoplay', False)
+        return str(paused).lower()